+++ R/map.R
@@ -38,11 +38,11 @@
 #'   map(~ rnorm(10, .x))
 #'
 #' # Extract by name or position
-#' # .null specifies value for elements that are missing or NULL
+#' # .default specifies value for elements that are missing or NULL
 #' l1 <- list(list(a = 1L), list(a = NULL, b = 2L), list(b = 3L))
-#' l1 %>% map("a", .null = "???")
-#' l1 %>% map_int("b", .null = NA)
-#' l1 %>% map_int(2, .null = NA)
+#' l1 %>% map("a", .default = "???")
+#' l1 %>% map_int("b", .default = NA)
+#' l1 %>% map_int(2, .default = NA)
 #'
 #' # Supply multiple values to index deeply into a list
 #' l2 <- list(
@@ -53,9 +53,9 @@
 #' l2 %>% map(c(2, 2))
 #'
 #' # Use a list to build an extractor that mixes numeric indices and names,
-#' # and .null to provide a default value if the element does not exist
+#' # and .default to provide a default value if the element does not exist
 #' l2 %>% map(list("num", 3))
-#' l2 %>% map_int(list("num", 3), .null = NA)
+#' l2 %>% map_int(list("num", 3), .default = NA)
 #'
 #' # A more realistic example: split a data frame into pieces, fit a
 #' # model to each piece, summarise and extract R^2